The Start
Created a Pill that will be the player and linked the camera to it. Created a platform for the pill to stand on that will act as the floor temporarily. Finally created player script that contains movement code such as looking with the camera and moving with WASD. The looking script does make the players camera become inverted but will fix in the next commit.